refactor: only keep function pointers as part of the AatTree, not AatNode
It is unnessary to keep the same function pointers that handle generic type behaviors in both AatTree and AatNode.
AatTree will make sure key and value are correct processed before inserting into AatNode. Doing so effectively
saved the memory needed to store a copy of all the function pointers in each AatNode.